+++ packages/vue3/package.json
@@ -71,11 +71,7 @@
     "parserOptions": {
       "ecmaVersion": 2020
     },
-    "rules": {
-      "no-empty-function": "off",
-      "no-empty-pattern": "off",
-      "@typescript-eslint/no-empty-function": ["off"]
-    },
+    "rules": {},
     "overrides": [
       {
         "files": [
