+++ testUtils/testWithMockedIntersectionObserver.ts
@@ -1,12 +1,21 @@
-function testWithMockedIntersectionObserver(cb: (...args:any) => void) {
-  let intersectCallback: (root?: Element, rootMargin?: string, thresholds?: number) => void = () => {}; // will be
+export const testWithMockedIntersectionObserver = (
+  cb: (...args: any) => void
+) => {
+  let intersectCallback: (
+    root?: Element,
+    rootMargin?: string,
+    thresholds?: number
+  ) => void = () => {}; // will be
   // populated later
-  let nativeIntersectionObserver = global.IntersectionObserver;
+  const nativeIntersectionObserver = global.IntersectionObserver;
 
   // Mock IntersectionObserver
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   global.IntersectionObserver = class {
-    constructor(cb: (root?: Element, rootMargin?: string, thresholds?: number) => any) {
+    constructor(
+      cb: (root?: Element, rootMargin?: string, thresholds?: number) => any
+    ) {
       // This is the callback that notifies when an intersection occurs
       // We'll store it to use it later
       intersectCallback = cb;
@@ -20,6 +29,4 @@ function testWithMockedIntersectionObserver(cb: (...args:any) => void) {
 
   // restore
   global.IntersectionObserver = nativeIntersectionObserver;
-}
-
-export default testWithMockedIntersectionObserver;
+};
